Add files via upload
An application for managing a shopping mall was developed using Assembly Language Programming in 8086Emulator. The goal of this project is to replicate a mall with a variety of clothes and accessory stores. The user can access the store of his favorite brand using the application and choose the products of his choosing. Along with the names of the clothing goods, a pricing list is also displayed. The user gets the option to choose the item and the number of goods they want to buy. The final output of the application is the entire bill. Therefore, the purpose of this project is to reinvent the way in which a person shops for goods in a mall.
With the use of assembly language programming, we hope to build a useful application with this project. ALP is usually viewed as being outmoded in favor of other languages like C, C++, Python, and others. We anticipate that leveraging 8086 emulators to create a full mall, will rekindle interest in and draw people to this language. Making a management system for a shopping center also has great value because it will allow customers to shop and swiftly create their bills.